refactor(ui): extract shared filter primitives and webhook delivery renderers
Lifts two sets of building blocks out of their current homes so a second
consumer can use them, with no behaviour change:

- components/filters/primitives.tsx takes FilterSection,
  CheckboxFilterItem, SearchableCheckboxList and useAutoFocusOnOpen out of
  mcpFilterSidebar.tsx, where they had been hand-copied from
  logsFilterSidebar.tsx. mcpFilterSidebar now imports them (-207 lines).
  logsFilterSidebar still holds its own copy; migrating it is a follow-up.

- webhooks/views/deliveries.utils.tsx takes the outcome/status-code badge
  renderers, the attempt-sequence chips and the two-level groupDeliveries
  helper out of webhookDetailsSheet.tsx. The sheet imports them and is
  otherwise untouched.

Both moves are verbatim; the sheet's delivery history renders exactly as
before.
